# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UMKM Go API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended]
        - name: tier
          in: query
          schema:
            type: string
            enum: [free, pro, enterprise]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{id}/suspend:
    post:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/unsuspend:
    post:
      operationId: unsuspendUser
      tags: [admin]
      summary: Unsuspend a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User unsuspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/revenue:
    get:
      operationId: getAdminRevenue
      tags: [admin]
      summary: Get revenue data (admin)
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueData"
...
  /admin/logs:
    get:
      operationId: getAuditLogs
      tags: [admin]
      summary: Get audit logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: action
          in: query
          schema:
            type: string
        - name: actor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
...
  /flags:
    get:
      operationId: listFeatureFlags
      tags: [flags]
      summary: List feature flags (admin)
      responses:
        "200":
          description: Feature flags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureFlag"
    post:
      operationId: createFeatureFlag
      tags: [flags]
      summary: Create feature flag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeatureFlagRequest"
      responses:
        "201":
          description: Feature flag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlag"
  /flags/{id}:
    put:
      operationId: updateFeatureFlag
      tags: [flags]
      summary: Update feature flag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeatureFlagRequest"
      responses:
        "200":
          description: Updated flag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlag"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
        tier:
          type: string
          enum: [free, pro, enterprise]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
          format: date-time
        hasStore:
          type: boolean
      required:
        - id
        - name
        - email
        - role
        - tier
        - status
        - createdAt
        - hasStore
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        storeName:
          type: string
        tier:
          type: string
          enum: [free, pro, enterprise]
        status:
          type: string
          enum: [active, suspended]
        revenue:
          type: number
        joinDate:
          type: string
          format: date-time
        productCount:
          type: integer
      required:
        - id
        - name
        - email
        - tier
        - status
        - revenue
        - joinDate
    AdminUserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    RevenueData:
      type: object
      properties:
        totalRevenue:
          type: number
        mrr:
          type: number
        mrrGrowth:
          type: number
        paymentSuccessRate:
          type: number
        avgOrderValue:
          type: number
        trend:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              revenue:
                type: number
            required:
              - date
              - revenue
        byTier:
          type: object
          properties:
            free:
              type: number
            pro:
              type: number
            enterprise:
              type: number
          required:
            - free
            - pro
            - enterprise
        recentTransactions:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              userEmail:
                type: string
              amount:
                type: number
              tier:
                type: string
              status:
                type: string
              date:
                type: string
            required:
              - id
              - userEmail
              - amount
              - tier
              - status
              - date
      required:
        - totalRevenue
        - mrr
        - mrrGrowth
        - paymentSuccessRate
        - avgOrderValue
        - trend
        - byTier
        - recentTransactions
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
        actor:
          type: string
        action:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
        status:
          type: string
          enum: [success, failed]
        details:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - actor
        - action
        - resourceType
        - status
        - createdAt
    AuditLogListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...
    FeatureFlag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: boolean
        rolloutPercent:
          type: integer
        tier:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - rolloutPercent
        - createdAt
    CreateFeatureFlagRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: boolean
        rolloutPercent:
          type: integer
        tier:
          type: string
        description:
          type: string
      required:
        - name
    UpdateFeatureFlagRequest:
      type: object
      properties:
        status:
          type: boolean
        rolloutPercent:
          type: integer
        tier:
          type: string
        description:
          type: string